# nepflow/src/main/java/com/nepflow/NepenthesManagement/Controller/MultiHybridApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.nepflow.NepenthesManagement.Controller;

import com.nepflow.NepenthesManagement.Dto.LabelClonesDTO;
import com.nepflow.NepenthesManagement.Dto.LabelDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2024-06-11T01:56:15.806697200+02:00[Europe/Berlin]")
@Validated
@Tag(name = "MultiHybrid", description = "the MultiHybrid API")
public interface MultiHybridApi {

    default MultiHybridApiDelegate getDelegate() {
        return new MultiHybridApiDelegate() {};
    }

    /**
     * GET /clone/multi-hybrid/ : get all Multi Hybrids
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridGet",
        summary = "get all Multi Hybrids",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LabelDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/multi-hybrid/",
        produces = { "application/json" }
    )
    default ResponseEntity<List<LabelDTO>> cloneMultiHybridGet(
        
    ) {
        return getDelegate().cloneMultiHybridGet();
    }


    /**
     * GET /clone/multi-hybrid/ic : get all IC Multi Hybrid Clones
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridIcGet",
        summary = "get all IC Multi Hybrid Clones",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LabelClonesDTO.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/multi-hybrid/ic",
        produces = { "application/json" }
    )
    default ResponseEntity<List<LabelClonesDTO>> cloneMultiHybridIcGet(
        
    ) {
        return getDelegate().cloneMultiHybridIcGet();
    }


    /**
     * GET /clone/multi-hybrid/ic/{name} : get all IC Clones of a Multi Hybrid
     *
     * @param name  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridIcNameGet",
        summary = "get all IC Clones of a Multi Hybrid",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/multi-hybrid/ic/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridIcNameGet(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name
    ) {
        return getDelegate().cloneMultiHybridIcNameGet(name);
    }


    /**
     * POST /clone/multi-hybrid/ic/{name} : add a new iv clone to a Multi Hybrid
     *
     * @param name  (required)
     * @param labelClonesDTO Clone DTO with new values (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridIcNamePost",
        summary = "add a new iv clone to a Multi Hybrid",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/clone/multi-hybrid/ic/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridIcNamePost(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "Clone DTO with new values", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneMultiHybridIcNamePost(name, labelClonesDTO);
    }


    /**
     * PUT /clone/multi-hybrid/ic/{name}
     *
     * @param name  (required)
     * @param labelClonesDTO  (required)
     * @return OK (status code 200)
     *         or No nepenthes found (status code 404)
     */
    @Operation(
        operationId = "cloneMultiHybridIcNamePut",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "No nepenthes found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/clone/multi-hybrid/ic/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridIcNamePut(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneMultiHybridIcNamePut(name, labelClonesDTO);
    }


    /**
     * GET /clone/multi-hybrid/iv/{name} : get all Multi Hybrid IV Clones
     *
     * @param name  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridIvNameGet",
        summary = "get all Multi Hybrid IV Clones",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/multi-hybrid/iv/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridIvNameGet(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name
    ) {
        return getDelegate().cloneMultiHybridIvNameGet(name);
    }


    /**
     * POST /clone/multi-hybrid/iv/{name} : add a new iv clone to a Multi Hybrid
     *
     * @param name  (required)
     * @param labelClonesDTO Clone DTO with new values (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridIvNamePost",
        summary = "add a new iv clone to a Multi Hybrid",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/clone/multi-hybrid/iv/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridIvNamePost(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "Clone DTO with new values", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneMultiHybridIvNamePost(name, labelClonesDTO);
    }


    /**
     * PUT /clone/multi-hybrid/iv/{name}
     *
     * @param name  (required)
     * @param labelClonesDTO  (required)
     * @return OK (status code 200)
     *         or No nepenthes found (status code 404)
     */
    @Operation(
        operationId = "cloneMultiHybridIvNamePut",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "No nepenthes found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/clone/multi-hybrid/iv/{name}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridIvNamePut(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Parameter(name = "LabelClonesDTO", description = "", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneMultiHybridIvNamePut(name, labelClonesDTO);
    }


    /**
     * GET /clone/multi-hybrid/{name} : get all Multi Hybrid Clones of a Multi Hybrid
     *
     * @param name  (required)
     * @return OK (status code 200)
     *         or nepenthes not found (status code 404)
     */
    @Operation(
        operationId = "cloneMultiHybridNameGet",
        summary = "get all Multi Hybrid Clones of a Multi Hybrid",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "nepenthes not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/clone/multi-hybrid/{name}",
        produces = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridNameGet(
        @Parameter(name = "name", description = "", required = true, in = ParameterIn.PATH) @PathVariable("name") String name
    ) {
        return getDelegate().cloneMultiHybridNameGet(name);
    }


    /**
     * POST /clone/multi-hybrid/ : add a new Multi Hybrids
     *
     * @param labelClonesDTO  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "cloneMultiHybridPost",
        summary = "add a new Multi Hybrids",
        tags = { "MultiHybrid" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = LabelClonesDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/clone/multi-hybrid/",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<LabelClonesDTO> cloneMultiHybridPost(
        @Parameter(name = "LabelClonesDTO", description = "", required = true) @Valid @RequestBody LabelClonesDTO labelClonesDTO
    ) {
        return getDelegate().cloneMultiHybridPost(labelClonesDTO);
    }

}