# nepflow/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: API for Nepenthes Operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Nepflow Nepenthes
  version: 1.0.0
servers:
- url: http://localhost:8080/api
...
paths:
  /location:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDTO'
        description: LocationdTo with all necessary Attributes
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDTO'
          description: OK
        "500":
          description: Internal Error
      summary: add a new Location
      tags:
      - NepenthesMetadata
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: NepenthesMetadata
  /producer:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerDTO'
        description: DTO with all necessary Attributes
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProducerDTO'
          description: OK
        "500":
          description: Internal Error
      summary: add a new Producer
      tags:
      - NepenthesMetadata
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: NepenthesMetadata
...
  /clone/multi-hybrid/:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  allOf:
                  - $ref: '#/components/schemas/LabelDTO'
                type: array
          description: OK
      summary: get all Multi Hybrids
      tags:
      - MultiHybrid
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: add a new Multi Hybrids
      tags:
      - MultiHybrid
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
  /clone/multi-hybrid/{name}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
        "404":
          description: nepenthes not found
      summary: get all Multi Hybrid Clones of a Multi Hybrid
      tags:
      - MultiHybrid
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
  /clone/multi-hybrid/iv/{name}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: get all Multi Hybrid IV Clones
      tags:
      - MultiHybrid
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
    post:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        description: Clone DTO with new values
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: add a new iv clone to a Multi Hybrid
      tags:
      - MultiHybrid
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
    put:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
        "404":
          description: No nepenthes found
      tags:
      - MultiHybrid
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
...
  /clone/multi-hybrid/ic/{name}:
    get:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: get all IC Clones of a Multi Hybrid
      tags:
      - MultiHybrid
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
    post:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        description: Clone DTO with new values
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
      summary: add a new iv clone to a Multi Hybrid
      tags:
      - MultiHybrid
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
    put:
      parameters:
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabelClonesDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelClonesDTO'
          description: OK
        "404":
          description: No nepenthes found
      tags:
      - MultiHybrid
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: MultiHybrid
components:
  schemas:
    LocationDTO:
      type: string
    ProducerDTO:
      type: string
    LabelDTO:
      example:
        prefix: prefix
        name: name
      properties:
        name:
          type: string
        prefix:
          type: string
      type: object
    CloneDTO:
      properties:
        cloneId:
          type: string
        internalcloneId:
          type: string
        sex:
          type: string
        grex:
          $ref: '#/components/schemas/CloneGrexDTO'
        producer:
          type: string
        Location:
          type: string
      type: object
    HybridCloneDTO:
      allOf:
      - $ref: '#/components/schemas/CloneDTO'
      - properties:
          fatherName:
            type: string
          motherName:
            type: string
        type: object
    CloneGrexDTO:
      properties:
        cloneId:
          type: string
        father:
          $ref: '#/components/schemas/CloneGrexDTO_father'
        mother:
          $ref: '#/components/schemas/CloneGrexDTO_father'
      type: object
    LabelClonesDTO:
      example:
        clones:
        - null
        - null
        label:
          prefix: prefix
          name: name
      properties:
        label:
          $ref: '#/components/schemas/LabelDTO'
        clones:
          items:
            $ref: '#/components/schemas/LabelClonesDTO_clones_inner'
          type: array
      type: object
    CloneGrexDTO_father:
      oneOf:
      - $ref: '#/components/schemas/CloneDTO'
      - $ref: '#/components/schemas/HybridCloneDTO'
      x-one-of-name: CloneGrexDTOFather
    LabelClonesDTO_clones_inner:
      anyOf:
      - $ref: '#/components/schemas/CloneDTO'
      - $ref: '#/components/schemas/HybridCloneDTO'